requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.4.0
django-monthfield==0.1.3
django-urls==1.1.3
djangorestframework==3.12.1
Faker==4.1.3
importlib-metadata==2.0.0
path==15.0.0
pathlib==1.0.1
Pillow==7.2.0
pkg-resources==0.0.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uWSGI==2.0.19.1
zipp==3.2.0
